requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pymupdf
python-dotenv
axios